requirements.txt
django=2.1.2
djangorestframework=3.9.0
django-money=0.14.3
fcm-django==0.2.19